:truck: Rename: Detach useCodemirror hook
